% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ttest.R
\name{doTTest}
\alias{doTTest}
\title{Title}
\usage{
doTTest(
  data_in,
  col1,
  col2,
  col3 = NA,
  label.col = "row.names",
  var.equal = TRUE,
  alternative = "two.sided",
  debug = FALSE
)
}
\arguments{
\item{debug}{}
}
\description{
Title
}
