% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epitope.R
\name{findEpitopes}
\alias{findEpitopes}
\title{find epitopes using t-test statistics}
\usage{
findEpitopes(
  probes,
  logfc,
  pvalues,
  mean.signal,
  probe_meta,
  lfc.threshold = 0,
  pvalue.threshold = 0.05,
  two.sided = FALSE
)
}
\arguments{
\item{two.sided}{}
}
\description{
find epitopes using t-test statistics
}
