% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epitope.R
\name{findEpitopesTTestProbe}
\alias{findEpitopesTTestProbe}
\title{Title}
\usage{
findEpitopesTTestProbe(
  ttest_result_probe,
  probe_meta,
  lfc.threshold,
  pvalue.threshold
)
}
\arguments{
\item{pvalue.threshold}{}
}
\description{
Title
}
