% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{getEpitopeID}
\alias{getEpitopeID}
\title{Make Epitope Id}
\usage{
getEpitopeID(protein, start, stop)
}
\arguments{
\item{protein}{protein of the epitope}

\item{start}{first probe start position}

\item{stop}{last probe start position}
}
\value{
epitope identifer string
}
\description{
Make Epitope Id
}
\examples{
getEpitopeID("A", 1, 3)
}
