% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{loadProbeMeta}
\alias{loadProbeMeta}
\title{Load the probe metadata}
\usage{
loadProbeMeta(file_name = "all_sequences_except_wi.tsv.gz")
}
\arguments{
\item{file_name}{input filename}
}
\value{
data.frame of meta data from probes
}
\description{
Load the probe metadata
}
