% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{loadSeqMat}
\alias{loadSeqMat}
\title{Load the sequence matrix}
\usage{
loadSeqMat(file_name = "df_stacked.tsv", col_name = "INTENSITY")
}
\arguments{
\item{file_name}{input filename}

\item{col_name}{signal column to use (Default INTENSITY)}
}
\description{
Load the sequence matrix
}
