% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ttest.R
\name{doTTest}
\alias{doTTest}
\title{Title}
\usage{
doTTest(
  data_in,
  col1,
  col2,
  col3 = NA,
  label.col = "row.names",
  var.equal = TRUE,
  alternative = "two.sided",
  debug = FALSE
)
}
\arguments{
\item{data_in}{input data.frame}

\item{col1}{vector of column names for 1st condition}

\item{col2}{vector of column names for 2nd condition}

\item{col3}{not used}

\item{label.col}{which column to use for rows}

\item{var.equal}{var.equal parameter for t.test call}

\item{alternative}{alternative parameter for t.test call}

\item{debug}{print debugging information}
}
\value{
data.frame of results
}
\description{
Title
}
